mech part repath (#7009)
Tears apart the overloaded `/obj/item/mecha_parts` into a few different
paths.

`/obj/item/vehicle_chassis` for the base part that things attach on
(this is the mech construction holder)
`/obj/item/vehicle_part` for things you attach to chassis
`/obj/item/vehicle_component` for integral components
`/obj/item/vehicle_module` for equipment

This is to prepare for further work on mecha as a system, and to get
some of the massive line changes out of the way.

Also:

- tears apart a few files that were just too long to be handled and
groups everything with their mechs
- erases some code reuse