Using the is_avoiding_ground() check: Flying no longer has you touching the ground. (#6991)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Moves a few checks if someone's flying/hovering to use the
is_avoiding_ground check. It also removes the slowdown of a tile if you
aren't actually stepping on it. This check does include more, like if
you are only riding on something that's lifting you up enough, but
unless it flies itself, it'll trigger what you may be avoiding, and in
turn, will be affected by it (i.E falling down the chasm, there is no
delay in my test, so no Yoshiing over a single tile.)
Draft for now to see if there are more places to check for it, and as
the balance change is big enough to warrant additional thoughts.

## Why It's Good For The Game

Flying people no longer leave footprints, nor touch the ground for water
or lava. Logically, they aren't even slowed down by the tiles. This
heavily buffs flying, at least the ignoring the slowdown part, so that
may be removed for the sake of balance. The rest should be more
flavorful, or just making sense.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
tweak: If you aren't touching the ground, don't interact with things
there.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: MarsMond <no@email>