Allows holospheres to be used with the hardsuit IIS Module (#7062)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

It's a relativly simple change that allows a holosphere in sphere form
to be used as the synthetic helping with the controls of it.

## Why It's Good For The Game

I think the system is almost never used, at most by proteans doing their
own suit. It currently requires either a full station AI, which would be
quite limited to be put in the suit, a brain, not sure if you could
bring any person for that, or a pAI, which I think is currently the only
realistic option, if rare enough. Allowing the spheres to integrate too
means an active player can just quickly switch into it, and being talked
to about it. This should help finding problems with it, and gives way
for the species to interact with the hardsuits directly, as currently,
they can't wear one.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: Holosphere's can now be integrated with a hardsuit through the IIS
module.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: MarsMond <no@email>