New planet : Lythios 43a - The sky planet (#6986)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

"A chilly wind starts blowling in your back whil you descend from a
shuttle. It whistle on the rusted metal base you are starting on, a
mighty sky rig... Indeed, all arround you is a Ocean of clouds, it
"waves" floating lightly in air. But you also know that the views is
only one side of a coin... This is paradise, and underneath, deep under
the clouds, is an Acidic hell... Welcome, on Lythios 43a."

This PR adds a new 6 Z level exploration map, like Miaphus and other POI
planets.

This planet have random generated POI like Miaphus, but also new
activities beyond exploring.
Indeed, the planet is home of the Voidline Shuttle Racing competions,
and comes with 3 race shuttles to compete. (A fourth one, the Arrowhead,
is on Nebula)

New planet, new foes. Flying lizards, "Dragoons", roam the skies. While
green one are peacefull, the Blue, Yellow and Red ones are too be
avoided....

Speaking of that, to defy their control in the sky, youll need probably
a fighters. Good think Pinances are arround, and crew can lauch them
from NT Outpost Hyades, a EX-SDF base.

## Why It's Good For The Game

-New destination, with new POI. Offers new place to explore, in the
harder side in some places and easier in others.
-Use of a game mechanic, fighters, that isn't used much (they are
usefull to explore arround.
-New out of spawn activity : Shuttle Race.

Still a WIP.

![image](https://github.com/user-attachments/assets/1c82219f-52fe-48c0-9ea3-9f42bfaf3020)