Science Voidsuit (Hazard Bypass Suit) (#7070)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
This PR adds a new voidsuit for the game, mainly for the science
department, to be on more equal grounds on most departments having a
personalized hardsuit for their area of work as well. This also adds a
new voidsuit section, in the code, specifically for the science
department as well, since they were lacking one.

Further PRs I'll attempt, will be adding it into the maps and cargo
orders, if this one gets approved for addition.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
I believe that science should be able to get their own specialized,
round start voidsuits, much as the same for the other departments,
(Medical, Security, Exploration, and Cargo). One being more suited for
their line of work of handling an array dangerous artifacts and
exploring the hazardous enviroments with the exploration team, if the
time calls for it, or general experimentation. It's a much less combat
capable voidsuit. More designed to avoid the danger all together, rather
than confront it head on. A good all rounder for science, if they don't
know truly know what they're getting themselves into.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

![image](https://github.com/user-attachments/assets/5c34692a-730c-4561-8d83-b65e2cc0d893)

![image](https://github.com/user-attachments/assets/6ba31bd5-cd5f-4473-9743-1ba33011a25e)


## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl: Kade
add:  added the Hazard Bypass Voidsuit
imageadd: added sprites for the Hazard Bypass Voidsuit
code: Added a void spacesuit section for science, as it was previously
lacking.

/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->