Allows leash laws to be enforced (#7067)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Despite the availability of collars, and the presence of dogs, there are
no leashes. This means you can't bring them to the indoor park, while
keeping them leashed to not bother other guests of the park. Nor can you
go outside, without the dog running of to chase the wildlife. This fixes
that.

## Why It's Good For The Game

April fools out of the way, it's mostly a port from
[Virgo](https://github.com/VOREStation/VOREStation/pull/16813), with a
few tweaks to make it work here, and some others to allow more
flexibility. Variable length to allow some VV shenanigans, tested with
the limited possibilities for the cable ones, and allowing simple dogs
to be leashed (which could be expanded to more passive simple animals).

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: Leashes
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: MarsMond <no@email>