Audits-the-icons-folder. Moves 30% of dmi's to Unusued. (#7103)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Of the 2,572 total files in the icons folder this moves 745 to the
Unused folder, mirroring the icons layout before they were moved. Thats
29% of all files in icons!

I went through every single file in the Icons folder and searched for
its path in the codebase. The process was to search for the full path
from icons/* onwards and if nothing was found for that file to search
the file name directly and check to see if its path is broken or being
used in a weird way. I've submitted a few errors I've found to Issues
and left those dmi's that would be used if they were fixed where they
are.

I think this has also revealed a few additional bugs, the doors and
species folders have dmi's in them that match the dmi's in their sibling
sub-folders but are never referenced. I assume these were forgotten
about and never added in the code.

This cleans up the icons folder significantly and brings it closer to
the ideal compliance described in its README. Since this is a really big
change I'll the maintainers figure out if they want to merge it or work
off of it.

I did a few tests during the process to make sure nothing exploded, the
compiler is happy and I didn't see anything running around in game that
was blatantly icon broken after cleaning everything up. However I am
only one guy so I can't reasonably confirm I didn't miss something.

I did not touch any icon states inside the dmis as thats an entire
project on its own. Nor did I rearrange any dmi files except for the
ones moved to unused.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Makes icons significantly easier to review and look at.
Ensures only used files are in the Icons folder outside Unused.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
refactor: refactored icons folder
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->