Murphy marvelous musings on remodeling (#6989)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->


Oh god Oh Fuck, I'm redoing most of the ship from the ground up.
Re-arranging most of the internal bits, squishing things closer
together. Shifting maints to the exterior of the hull.

This is a comprehensive overhaul of the entire ship, shifting everything
around and simplifying some areas. A true 'Mark II" design.

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

It makes the ship feel more homey and less empty.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
tweak: Tweak is not a strong enough word for what I'm doing. 
tweak: Shifted bar, pool, garden. 
tweak: Hangers moved to aft of deck 1. 
tweak: Security and science significantly redesigned.
tweak: engineering deck modified to allow for bar in the forward half of
the deck.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: silicons <2003111+silicons@users.noreply.github.com>
Co-authored-by: AlphaM01 <AlphaaMachh@gmail.com>
Co-authored-by: AlphaM01 <33434925+AlphaM01@users.noreply.github.com>