Set up files for all the existing away missions, remove jungle_temple
As far as I can tell jungle_temple is connected to nothing, it looks like its setup to spawn in templates but nothing references this file and its from 6 months ago in a commit. I'm going to archive it. If the creator wants to finish it they can pull it back out.