Finish up
Confirmed maps work. Zoo does not and has been removed. it needs a few mobs repaired and the map needs to be cut down a lot as the map subsystem refuses to load it due to size.

Otherwise this code works and the gateway is repaired. The other functions like calibrating and teleporting are fine and currently work.