Small gun fixes (#7104)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Potentially dirty fixes, although there is only one new var to stop the
revolver spinning to use the full magazine equals chamber logic of that,
which does seem like it has the intended effect for double-barrel
shotguns.
The crossbow attack_self returns early as it attempts to switch
firemodes, which is why it can't be drawn, object to be thrown was just
moved into the shooter.

## Why It's Good For The Game

Double barrel shotguns should actually load two shells, and crossbows
are currently spawned in, but don't work.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
fix: crossbows can be drawn
fix: double-barrel is actually two barrels
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: MarsMond <no@email>