Rewrites the R-UST guide not to tell you the exact wrong way to do it (#7142)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Right now it tells you to put one tritium and one deuterium into the TEG
and start pumping that in before you turn on the gyrotrons. This is, if
you do it slowly enough, literally a *guaranteed* way to blow up the
engine. It is not safe. Deuterium-only is safe. I don't even know if it
was ever this way.

## Why It's Good For The Game

the guide should not be a newb trap

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
fix: the hardcoded guide for the R-UST no longer encourages you to blow
it up
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->