gives holospheres the reset to slot and set size verbs (#7154)
## About The Pull Request
title

shapeshifter_reset_to_slot is now shapeshifter_reset_to_slot_generic
which has parameters for the duration, cooldown and text


## Why It's Good For The Game
someone asked for it and it feels like something they should be able to
do

## Changelog

:cl:
add: gives holospheres the reset to slot and set size verbs
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->