Renamed Atlas substations/breakers to be consistent (#7158)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Renames most of the NSB Atlas substations and breaker boxes to label
which floor they're on and generally be more consistent. This should
help managing RCON be a lot smoother. See screenshots of the new layout:

![image](https://github.com/user-attachments/assets/432b2f4d-65e2-456b-acfc-5ce875e5e5c7)

![image](https://github.com/user-attachments/assets/22e2570d-bcc8-4847-993f-3db2f346c695)

![image](https://github.com/user-attachments/assets/4b75e119-fd10-4c91-8fba-0c562f47f9cf)

![image](https://github.com/user-attachments/assets/53a8a2ef-b815-4894-af61-774be71c289b)

## Why It's Good For The Game

The old RCON labels were inconsistent and somewhat confusing. This helps
engineering players, especially new ones, learn the RCON system quicker.
To compare, this is what the breakers were like before:

![image](https://github.com/user-attachments/assets/82e2b80f-f92d-4426-b6a6-ebfbe0b06b82)

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
tweak: Adjusted the RCON labels of the substations and breaker boxes on
the NSB Atlas.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->