Drone matrix upgrades touchup (#7178)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Just a slight change to how the drone upgrades are applied. Mostly
ensures they get only applied once, and don't delete and re-add the
upgraded item. The potential for infinite growth remains, although it
would need admin intervention to remove the already applied upgrades
from the list. It still will allow it to be upgrades even if the drone
got edited before, why I removed the initial() call for it.

## Why It's Good For The Game

As fun as exponentially growing power storage is, and with enough time,
an extremely tanky drone, upgrades should only apply once. The speed
upgrade actually is an upgrade again, as it's no delay anymore, but
tiles per seconds, so the sign had to switch.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
fix: Actuator enhancement is actually an enhancement again.
tweak: drone upgrades only get applied once
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: MarsMond <no@email>