Small Ship map : NEV Strelka (#7025)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

_Somewhere on Centcom NDV Marksman, near a window_

-"Hey, Officer, here are the manifest of the last few months from crew 7
EX 39."
-"Huh. Thank you... Wow, is it me or they are getting less people ?"
-"Yes, I agree... Yet they are getting some big bas- Holly molly, is
that the Strelka ? Its getting repaired ? My grandpa loved that ship !"
-"Huhuh. We only need a smaller crew to man it..."
_Both are looking at a the crew manifest folder_
-**"Bingo."**

The Strelka is one of the oldest NT Vessel. Started out as a Luxury
Yacht for NT, to carry VIPs and celebrities. Served also as a blockade
runner during the corporate wars. Its last function was as a pathmaker
and obstacle buster for the Marksman fleet. With being one of the
oldest, its also one of its smallest, yet still can be used as a day to
day vessel, with... a lot of sacrifices.

The Vessel got four decks, like the Triumph or Endeavour, but is arround
only 80 Tiles long and 64 tiles wide. Its main focus was mainly confort
and luxury at first, with its departments being, for a lot of them, just
cabins with walls that got destroyed to make more room.

![nev
strelka](https://github.com/user-attachments/assets/4dbe5d39-9506-4e51-b3ff-6aa81281172d)

## Why It's Good For The Game

The map was made at first to test how small a citadel station can be. 
_But I got carried away :D_
There are not a lot of players playing during some time of the day, and
a request for a smaller map, that makes people interact more with each
other, has been requested. This map should do it. Its also takes a
direction that is different from the Atlas, Tether, Triumph and
Endeavour, aspect wise.

The map is also a test to apply the different ideas given from the
exploration departement debates that the server saw multiple times.
Here, the exploration depatement is actually rather open. All of the
crew can easely acces the excursion shuttle hangard. The gears of EMTs
and Explorators (and pathfy) are still here, but located in Science and
Medical (The Pathfy still get is own office). The goal is to have away
team with different crew of multiple departement more easely, without
the need to change access.
Since the vessel is a NEV, the pathfinder gets also acces to the bridge,
where he can assist with charter and helm... And of course, since this
vessel have a Obstacle removal Ballista, the pathfinder (and command)
can use it.

The main drawback of the vessel would be that some science activities
are missing, mostly those that can be found in the western plain outpost
of Lythios 43. So the remaining activities are : Robotics, RnD, Circuit,
Chemistry, Xenobiology, Xenobotany, Telescience, Xenoarchelogy, And
Anomaly.

Another drawback is the lack of maintenance tunnels. There are some
maintenance area, but were on other maps they are arround 30% to 45% of
a map, on this vessel, they be arround 10%. SMES are in glass boxes that
requires Engineering acces to open.

Other than that, all departments and activities are here. It is also
able to welcome traveler and trading shuttles in its Deck 4 hangar.

**It is still a WIP, and would require testing**

![StrongDMM-2025-03-03 16 32
27](https://github.com/user-attachments/assets/06cf0585-d9dc-4a84-9d2a-707ee6698f51)
![StrongDMM-2025-03-03 16 47
15](https://github.com/user-attachments/assets/3accc9a8-0273-442d-8b76-74f7c35c177f)
![StrongDMM-2025-03-03 16 47
34](https://github.com/user-attachments/assets/76b88d5d-3c25-44b4-b3b4-2acce63d0bc4)
![StrongDMM-2025-03-03 16 47
53](https://github.com/user-attachments/assets/1d2d7cd4-9f69-4f4a-9313-39449354b20d)

---------

Co-authored-by: LordME <58342752+TheLordME@users.noreply.github.com>