Shadekin-Spacekin-Futureproofing
Biggest change is that Space turf counts as dark for Shadekin, and those Shadekin who would lose energy in the light (Yellow, Purple, Orange) now instead have minimal regeneration in the light. This will be important for the future change to Shadekin with them travelling through space, since space is always bright.