Ghostrole : Haddi's Folley System Defence Force (#7217)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

PR needs a bit of work outside it (written SOP), advised not to merge or
TM yet (staff might need to speak about it first)
-------

"Soldiers, Officers, polish your boots, clean your ships and be prepared
for some new orders and reorganisation."

_Thats what we heard today. With the the massive migration of people on
Miaphus those past few month, life completly changed. We are far from
the few colonist and convincts mounting up their little governement to
rule and organise a few villages, and now the SDF needs to be able to
respond more quickly to threats in order to protect the system._

_New uniform arrived, and a few small outpost started popping up places
on Miaphus'irra. I heard that corporations injected a bit of funds ro
buy us new, smaller vessels, but whats truth or speculation is a bit
hard to tell on my position._

_I know that you are a bit scared for me, but it will be alright, I
Swear.
Your brother, Crewman Mawra'Kast_

--------

PR adds  : 
- 3 Uniform meant to be used by Haddis Folly Gouvernement Workers ,
Soldiers, and Officers.
_Those uniform are based on prison Jumpsuits, but refined to be more
offical looking. A reminded of the origin of this new nation to the
people wearing them_

![image](https://github.com/user-attachments/assets/5beb7918-3eff-49c8-8adf-f4dca2b73841)

![image](https://github.com/user-attachments/assets/403b4958-eabf-42c1-bbdd-6c9e5df0548e)

![image](https://github.com/user-attachments/assets/46dc762c-b1d9-42a7-8959-92ab3d4f3fe4)

- New ID

![image](https://github.com/user-attachments/assets/7dcf488c-153c-4328-8d4f-7066235b7461)


- A SDF Ghost role.
_Made of 5 slots, meant to have 1 to 2 informal officers, and max 4
soldiers.
Spawning on them are : uniform, standart radio, standart clothes.
Available to them are 4 service rifles (tajaran), 4 boot knifes, GPS, 2
officer pistols (Revolver and Deagle) headwear, standart space
voidsuits, 2 fighters (One gunship, One duke)
They have little to no medical gear._

- A shuttle (Corvette called SDF Singing shining knight) and a small
Base (Outpost 12)
The corvette is the main base. Ghost role spawn there. They get a
balista dispercer canon, 2 fighters (see above and on picture)


![image](https://github.com/user-attachments/assets/e9cfa2a6-b524-4c8c-af8a-08ceac82562a)

![image](https://github.com/user-attachments/assets/7ba9d1e4-6364-4455-ba64-b6cf495a424a)


## Why It's Good For The Game

This are the pros :

- Possibility to attract or bring back player that wish for more PVE
action, or a little more military RP.

- More Interaction too, The goal of this ghostrole is to be hunting bad
things that exploration team wouldn't dare approche, help the main NT
station or simply interact with them, giving medical (They will need
medical help) , cargo (They might request things), security (No way ISD
are sleeping to soldiers interacting with the Crew) , command (NO WAY
COMMAND ARE SLEEPING TO SOLDIERS INTERACTING WITH THE CREW), exploration
(No way exploration are sleeping to potential buddies to shoot bad guys
and critters with :D) and even service (No way soldiers are not walking
in a manned bar)

- Eventmin paradise. There is an actual SDF ship there, with Haddis
folly uniform and ID.

Now, the cons

- Like I said at the top of the PR description : 
This is a more militaristic role, that will require their own SOP
This is also a rather big shift to the server. Yes, indeed it opens up
new possibilities, and makes the system a tad more lively, but it can be
interpreted as a departure of what a part of the playerbase associate
Citadel with.
There might also be a slight griefer opportunity. Not only from "Sol
Bald man, emmissary of the curse of rah, wielder of the phoron tank and
welder", but also invonlotary from absolutly normal players that acts a
tad antagonistic.

Personnaly, once the SOP is put in place, and after hearing feedback and
correcting possible issue, I do thing it will contribute in more good
than bad for the server.