Ports sounds for writing on paper from baystation (#7282)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Ports https://github.com/Baystation12/Baystation12/pull/21269, which
adds two sound effects for when writing on paper:
https://freesound.org/people/Robinhood76/sounds/60825/
https://freesound.org/people/Robinhood76/sounds/257380/. Both these
sounds are CC-BY-NC.

Example video (sound is at 10 volume, I lowered it to 5 since it sounded
a little loud. This is just to show the concept in-game):


https://github.com/user-attachments/assets/fc333843-4715-445a-bb6b-461507ed17f8


## Why It's Good For The Game

Helps people know when you finish writing during paperwork RP.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
Add: Added sounds for writing on paper.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->