New POI : The Osiris debrisfield (#7137)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
_See
https://discord.com/channels/161574200581029888/857117946957266975/1274579475656347689
for the context_

**[OCULUM NEWS NETWORK - GALACTIC DISPATCH]**
_"Thank you Jerry for this fascinating report about the... Farwillas...
In other news...
The Hadii's Folly Governement, the local SDF, and multiple corporation
finaly relaxed the restriction arround a newly formed debrisfield. This
is an opportunity for a new wave of vessel scavengers, able to earn a
good deal. Report from Lrin'Lanri."_

_"2568-08-17 :_
_The BNV 'HELL-WALKER', a military vessel, hijacked by its Rogue AI
OSIRIS, Exploded.
The shockwave of the explosion destroyed every ship that wasn't shielded
enough.
A lot of factions lost souls : Civilians, Military, Corporate, Pirate...
	Worst of all : The casulty list is still incomplete._
	
_This is one of the biggest catastrophe of the the Lythios 43 system,
impacting the life of many. Months later, the shockwave still haunts a
lot of people, with many civilians on Miaphus and 43a recalling the
events. A debris field, 2 times bigger than the Lythios 43 first debris
field, had been restricted acces by the governement, the SDF, and
corporations such as Nanotrasen. A memorial station, build by the crew
of the NSV Von Braun and citizens of hadiis folly, was left serving as
the gates of this space graveyard._

_Yet, the SDFs, corporations and Governement declared lifting any travel
restrictions to the place, the debris field being judged stable enough.
A struck of luck for Explorers, but also a gold mine for Vessel
Scavengers, repairing and using or selling wrecks._

_Lets hope that their work will heal the spirits of the area."_

_"Thank you, Lrin'Lanri. Up next, Carp Sushi, a dish returning on your
pla..."_

## Why It's Good For The Game

Basically, The PR adds a New POI : 
The Osiris Debris Field.

The main goal of this POI is : 
* More POIs
* Concrete Aftermath of a truly great event in the universe, and visible
impact it had in space
* A new Salvaging mechanic for NT crew, but also traveler, with a new
alt title (Scavenger / Salvager)

So the PR at its current state adds
* Debris Field with new wrecks arround (Smaller Debris have been copy
and pasted from the normal Debris field (THE NORMAL DEBRIS FIELD WILL
STILL BE ARROUND)
* 2 vessel Wrecks that can be repaired (Colonial II, a luxury Liner -
Daidalos, a Merc Warship) and a Shuttle (Spacena Cargoravana, a mix
between the Caravan and Adventurer for those Spacena Shuttle fans). More
will be made.
~~*New Scavenger ghoste role, with a New Big Ship docked on Nebula, a
Mega Tug able to carry said Repairables Wrecks on a trailer.~~
**_The scavenger ghostrole will be added in a later PR wich will be
focused on more Offmaps jobs_**

**Memorial Station**

![image](https://github.com/user-attachments/assets/87cf604e-be99-42f8-951e-6ccc199103a2)

---------

Co-authored-by: Blubelle <57083662+BlueWildrose@users.noreply.github.com>